+++ packages/components/src/components/my-position/GeoLocationProvider.ts
@@ -1,10 +1,13 @@
+import { IPositionProvider } from '../../types/position-provider.interface';
+
 /**
  * GeoLocationProvider provides methods to access the device's location via the GeoLocation API ({@link https://developer.mozilla.org/en-US/docs/Web/API/Geolocation_API GeoLocation API}).
  */
-export class GeoLocationProvider {
+export class GeoLocationProvider implements IPositionProvider {
 
     /**
      * The current position of the device if received.
+     *
      * @type {GeolocationPosition | null}
      * @see {@link https://developer.mozilla.org/en-US/docs/Web/API/GeolocationPosition}
      */
@@ -18,17 +21,19 @@ export class GeoLocationProvider {
 
     /**
      * Check if it is possible to get the position of the device using the GeoLocation API.
+     *
      * @returns {boolean} True if the GeoLocation API is available in the browser, false otherwise.
      */
-    public static isAvailable(): boolean {
+    public isAvailable(): boolean {
         return 'geolocation' in navigator;
     }
 
     /**
      * Check if the user has already granted permission to access the device's location.
+     *
      * @returns {Promise<boolean>} A promise that resolves to true if permission is granted, false otherwise.
      */
-    public static async isAlreadyGranted(): Promise<boolean> {
+    public async isAlreadyGranted(): Promise<boolean> {
         if ('permissions' in navigator === false || 'query' in navigator.permissions === false) {
             return Promise.resolve(false);
         } else {
@@ -39,13 +44,14 @@ export class GeoLocationProvider {
 
     /**
      * Listen for position updates from the GeoLocation API.
+     *
      * @param {number} maxAccuracy - The maximum accuracy of the position to accept.
      * @param {Function} positionError - Callback function to call when an error occurs while getting the position.
      * @param {Function} positionInaccurate - Callback function to call when the position is inaccurate.
      * @param {Function} positionRequesting - Callback function to call when the position is being requested.
      * @param {Function} positionReceived - Callback function to call when a position is received.
      */
-    public static listenForPosition(
+    public listenForPosition(
         maxAccuracy: number,
         positionError: (error?: GeolocationPositionError) => void,
         positionInaccurate: (accuracy: number) => void,
@@ -59,21 +65,21 @@ export class GeoLocationProvider {
 
         positionRequesting();
         const requestTime = Date.now();
-        this.positionWatchId = navigator.geolocation.watchPosition(
+        GeoLocationProvider.positionWatchId = navigator.geolocation.watchPosition(
             position => {
                 // If position is the same as before, don't do anything.
                 if (
-                    this.currentPosition
-                    && position.coords.accuracy === this.currentPosition.coords.accuracy
-                    && position.coords.latitude === this.currentPosition.coords.latitude
-                    && position.coords.longitude === this.currentPosition.coords.longitude
+                    GeoLocationProvider.currentPosition
+                    && position.coords.accuracy === GeoLocationProvider.currentPosition.coords.accuracy
+                    && position.coords.latitude === GeoLocationProvider.currentPosition.coords.latitude
+                    && position.coords.longitude === GeoLocationProvider.currentPosition.coords.longitude
                 ) {
                     return;
                 }
 
-                this.currentPosition = position;
+                GeoLocationProvider.currentPosition = position;
 
-                if (!(this.currentPosition.coords.accuracy <= maxAccuracy)) {
+                if (!(GeoLocationProvider.currentPosition.coords.accuracy <= maxAccuracy)) {
                     positionInaccurate(position.coords.accuracy);
                 } else {
                     positionReceived(position);
@@ -82,7 +88,7 @@ export class GeoLocationProvider {
             error => {
                 // Firefox may throw both success and a timeout error (https://bugzilla.mozilla.org/show_bug.cgi?id=1283563).
                 // We mitigate this by not passing on error if a correct position has been given since asking for it.
-                if (error.code !== 3 || !this.currentPosition || this.currentPosition.timestamp <= requestTime) {
+                if (error.code !== 3 || !GeoLocationProvider.currentPosition || GeoLocationProvider.currentPosition.timestamp <= requestTime) {
                     positionError(error);
                 }
             }
@@ -95,11 +101,11 @@ export class GeoLocationProvider {
      *
      * @returns {void}
      */
-    public static stopListeningForPosition(): void {
-        if (this.positionWatchId !== null) {
-            navigator.geolocation.clearWatch(this.positionWatchId);
-            this.positionWatchId = null;
+    public stopListeningForPosition(): void {
+        if (GeoLocationProvider.positionWatchId !== null) {
+            navigator.geolocation.clearWatch(GeoLocationProvider.positionWatchId);
+            GeoLocationProvider.positionWatchId = null;
         }
-        this.currentPosition = null;
+        GeoLocationProvider.currentPosition = null;
     }
 }