+++ packages/react-components/src/components/MIMap/MIMap.jsx
@@ -25,7 +25,8 @@ MIMap.propTypes = {
     resetUICounter: PropTypes.number,
     mapOptions: PropTypes.object,
     onInitialized: PropTypes.func,
-    gmMapId: PropTypes.string
+    gmMapId: PropTypes.string,
+    devicePosition: PropTypes.object
 }
 
 /**
@@ -45,8 +46,9 @@ MIMap.propTypes = {
  * @param {function} [props.onInitialized] - Callback for when the MapsIndoors instance (https://app.mapsindoors.com/mapsindoors/js/sdk/latest/docs/mapsindoors.MapsIndoors.html)
  *    and position control and knowledge of the View mode switch is ready. The instance, position control and boolean for if the view mode switch is active is given as payload.
  * @param {string} [props.gmMapId] - The Google Maps Map ID for custom styling.
+ * @param {object} [props.devicePosition] - Device position object with coords and timestamp for custom positioning.
  */
-function MIMap({ apiKey, gmApiKey, mapboxAccessToken, center, zoom, bounds, bearing, pitch, resetUICounter, mapOptions, onInitialized, gmMapId }) {
+function MIMap({ apiKey, gmApiKey, mapboxAccessToken, center, zoom, bounds, bearing, pitch, resetUICounter, mapOptions, onInitialized, gmMapId, devicePosition }) {
 
     const [mapType, setMapType] = useState();
     const [mapsIndoorsInstance, setMapsIndoorsInstance] = useState();
@@ -191,6 +193,7 @@ function MIMap({ apiKey, gmApiKey, mapboxAccessToken, center, zoom, bounds, bear
                 mapInstance={mapViewInstance}
                 onPositionControl={setPositionControl}
                 brandingColor={mapOptions?.brandingColor}
+                devicePosition={devicePosition}
             />
         )}
     </>